requirements/ci.txt
-r base.txt

coverage==5.5
factory-boy==3.2.0
django_mutpy==0.1.2
